requirements.txt
python-decouple
django==1.7.4
pytz